Fix terminal wheel fallback and Ghostty display offsets
Preserve printable-text fallback for unmapped keystrokes and align Ghostty snapshot cell and damage lines with the current display offset.

+++ crates/hunk-terminal/src/snapshot.rs
@@ -82,7 +82,7 @@ pub struct TerminalModeSnapshot {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct TerminalDamageLineSnapshot {
-    pub line: usize,
+    pub line: i32,
     pub left: usize,
     pub right: usize,
 }
